+++ src/ffi/client.rs
@@ -1,8 +1,7 @@
+use std::ffi::c_int;
 use std::pin::Pin;
 use std::ptr;
 use std::sync::Arc;
-
-use libc::c_int;
 
 use crate::client::conn;
 use crate::rt::Executor as _;